add elementpul delete modal

+++ src/views/settings/DefaultMessage.vue
@@ -118,15 +118,11 @@
 <script>
 import axios from "@/gateway/backendapi";
 import messageOptions from '../../services/defaultmessage/default_message_service';
-import Toast from 'primevue/toast';
-import ConfirmDialog from 'primevue/confirmdialog';
 import Tooltip from 'primevue/tooltip';
 import finish from '../../services/progressbar/progress'
+import { ElMessage, ElMessageBox } from 'element-plus'
 export default {
   components:{
-    Toast,
-    ConfirmDialog
-
   },
     directives: {
     'tooltip': Tooltip
@@ -143,27 +139,33 @@ export default {
   },
   methods: {
        deletePop(id) {
-        this.deleteDefaultmessage(id)
-            // this.$confirm.require({
-            //     message: 'Are you sure you want to Delete?',
-            //     header: 'Delete Confirmation',
-            //     icon: 'pi pi-exclamation-circle',
-            //     acceptClass: 'confirm-delete',
-            //     rejectClass: 'cancel-delete',
-            //     accept: () => {
-            //       this.deleteDefaultmessage(id)
-            //         //callback to execute when user confirms the action
-            //     },
-            //     reject: () => {
-            //         'No internet'
-            //     }
-            // });
+        ElMessageBox.confirm(
+        'Are you sure you want to proceed?',
+        'Warning',
+        {
+          confirmButtonText: 'OK',
+          cancelButtonText: 'Cancel',
+          type: 'warning',
+        }
+      )
+        .then(() => {
+          this.deleteDefaultmessage(id)
+        })
+        .catch(() => {
+          ElMessage({
+            type: 'info',
+            message: 'Delete canceled',
+          })
+        })
         },
     async deleteDefaultmessage(id){
       try {
         await axios.delete('/api/Settings/DeleteDefaultMessage/'+id);
         this.defaultMessage = this.defaultMessage.filter(i => i.id !== id);
-         this.$toast.add({severity:'success', summary: '', detail:'Message Deleted Successfully', life: 3000});
+        ElMessage({
+              type: 'success',
+              message: 'Message Deleted Successfully',
+            })
       } catch (error){
         finish()
         console.log(error);
